Documentation Event Handling.
