+++ src/rust/ide/view/graph-editor/src/component/visualization/foreign/java_script/instance.rs
@@ -229,19 +229,8 @@ impl Instance {
                     frp.data_receive_error.emit(Some(e));
                 }
             });
-
-            let mouse_up       =  scene.mouse.frp.up.clone_ref();
-            let mouse_down     =  scene.mouse.frp.down.clone_ref();
-            let mouse_wheel    =  scene.mouse.frp.wheel.clone_ref();
-            let mouse_position =  scene.mouse.frp.position.clone_ref();
-            let keyboard_up    =  scene.keyboard.frp.up.clone_ref();
-            let keyboard_down  =  scene.keyboard.frp.down.clone_ref();
-            caught_mouse       <- any_(mouse_up,mouse_down,mouse_wheel,mouse_position);
-            caught_keyboard    <- any_(keyboard_up,keyboard_down);
-            caught_event       <- any(caught_mouse,caught_keyboard);
-            should_process     <- caught_event.gate(&frp.is_active);
-            eval_ should_process (scene.current_js_event.pass_to_dom.emit(()));
         }
+        frp.pass_events_to_dom_if_active(scene,network);
         self
     }
